+++ demo/pg_demo.go
@@ -3,7 +3,7 @@ package demo
 import (
 	"fmt"
 	"os"
-	"time"
+	"path/filepath"
 
 	"github.com/anynines/a9s-cli-v2/makeup"
 )
@@ -21,7 +21,8 @@ const demoA8sDeploymentLocalDir = "a8s-deployment"
 const certManagerNamespace = "cert-manager"
 const certManagerManifestUrl = "https://github.com/cert-manager/cert-manager/releases/download/v1.12.0/cert-manager.yaml"
 const defaultDemoSpace = "default"
-const systemName = "a8s Postgres control plane"
+const A8sSystemName = "a8s Postgres Control Plane"
+const A8sSystemNamespace = "a8s-system"
 
 var BackupInfrastructureProvider string // e.g. AWS
 var BackupInfrastructureRegion string   // e.g. us-east-1
@@ -96,47 +97,103 @@ func CheckPrerequisites() {
 	}
 }
 
-func WaitForSystemToBecomeReady() {
-	makeup.PrintH1("Waiting for the " + systemName + " to become ready...")
+/*
+Applies the manifests of the a8s-deployment repository and thus installs a8s PG.
+*/
+func ApplyA8sManifests() {
+	makeup.PrintH1("Applying the a8s Data Service manifests...")
+	kustomizePath := filepath.Join(DemoConfig.WorkingDir, demoA8sDeploymentLocalDir, "deploy", "a8s", "manifests")
+	KubectlApplyKustomize(kustomizePath)
+	makeup.PrintCheckmark("Done applying a8s manifests.")
+}
 
-	allGood := true
+/*
+Represents the state of a Pod which is expected to be running at some point.
+The attribute "Running" is meant to be updated by a control loop.
+*/
+type PodExpectationState struct {
+	Name    string
+	Running bool
+}
 
-	//TODO Make configurable or move to beginning of file for better maintainability
-	expectedPodPrefixes := []struct {
-		name    string
-		running bool
-	}{
-		{"a8s-backup-controller-manager", false},
-		{"postgresql-controller-manager", false},
-		{"service-binding-controller-manager", false},
+func WaitForA8sSystemToBecomeReady() {
+	expectedPods := []PodExpectationState{
+		{Name: "a8s-backup-controller-manager", Running: false},
+		{Name: "postgresql-controller-manager", Running: false},
+		{Name: "service-binding-controller-manager", Running: false},
 	}
-	systemNamespace := "a8s-system"
 
-out:
-	for {
-		// We start optimistically that all pods are running
-		allGood = true
-		for _, expectedPodPrefix := range expectedPodPrefixes {
-			makeup.Print("Checking the " + expectedPodPrefix.name + "...")
-			if checkIfPodHasStatusRunningInNamespace(expectedPodPrefix.name, systemNamespace) {
-				makeup.PrintCheckmark("The " + expectedPodPrefix.name + " appears to be running.")
-				expectedPodPrefix.running = true
-			} else {
-				// Sadly, at least one pod isn't running so we need another loop iteration
-				makeup.PrintFail("The " + expectedPodPrefix.name + " is not ready (yet).")
-				allGood = false
-			}
+	WaitForSystemToBecomeReady(A8sSystemNamespace, A8sSystemName, expectedPods)
+}
 
-			if allGood {
-				makeup.PrintSuccessSummary("The " + systemName + " appears to be ready. All expected pods are running.")
-				break out
-			} else {
-				makeup.PrintWait("The " + systemNamespace + " is not ready (yet), let's try again in 5s ...")
-				time.Sleep(5 * time.Second)
-			}
+func WaitForServiceInstanceToBecomeReady(namespace, serviceInstanceName string, nrOfInstances int) {
+	expectedPods := make([]PodExpectationState, 3)
+
+	for i := 0; i < nrOfInstances; i++ {
+		expectedPods[i] = PodExpectationState{
+			Name:    fmt.Sprintf("%s-%d", serviceInstanceName, i),
+			Running: false,
 		}
 	}
-	makeup.WaitForUser(UnattendedMode)
+
+	WaitForSystemToBecomeReady(namespace, serviceInstanceName, expectedPods)
+}
+
+func CreatePGServiceInstance() {
+	makeup.PrintH1("Creating a a8s Postgres Service Instance...")
+
+	EstablishConfigFilePath()
+
+	if !LoadConfig() {
+		makeup.ExitDueToFatalError(nil, "There is no config, yet. Please create a demo environment before attempting to create a service instance.")
+	}
+
+	makeup.Print("Using default values for creating the instance.")
+
+	// Stage 1: apply static manifest
+	// TODO stage 2: create struct, generate manifest based on parameters
+
+	exampleManifestPath := filepath.Join(A8sDeploymentExamplesPath(), "postgresql-instance.yaml")
+
+	makeup.PrintInfo("The YAML manifest of the service instance is located at: " + exampleManifestPath)
+
+	makeup.Print("The YAML manifest contains: ")
+	err := makeup.PrintYAMLFile(exampleManifestPath)
+
+	if err != nil {
+		makeup.ExitDueToFatalError(err, "Can't read service instance manifest from "+exampleManifestPath)
+	}
+
+	KubectlApplyF(exampleManifestPath)
+}
+
+// Refactor to DRY with Create ... > CRUDPGServiceInstance
+func DeletePGServiceInstance() {
+	makeup.PrintH1("Deleting a a8s Postgres Service Instance...")
+
+	EstablishConfigFilePath()
+
+	if !LoadConfig() {
+		makeup.ExitDueToFatalError(nil, "There is no config, yet. Please create a demo environment before attempting to create a service instance.")
+	}
+
+	makeup.Print("Using default values for deleting the instance.")
+
+	// Stage 1: apply static manifest
+	// TODO stage 2: create struct, generate manifest based on parameters
+
+	exampleManifestPath := filepath.Join(A8sDeploymentExamplesPath(), "postgresql-instance.yaml")
+
+	makeup.PrintInfo("The YAML manifest of the service instance is located at: " + exampleManifestPath)
+
+	makeup.Print("The YAML manifest contains: ")
+	err := makeup.PrintYAMLFile(exampleManifestPath)
+
+	if err != nil {
+		makeup.ExitDueToFatalError(err, "Can't read service instance manifest from "+exampleManifestPath)
+	}
+
+	KubectlDeleteF(exampleManifestPath)
 }
 
 func PrintDemoSummary() {