+++ Changelog.md
@@ -5,6 +5,9 @@
 * Now also the `a8s-demo` repository with its demo application manifests is cloned into {workdir}.
     * In theory the `a8s-demo` repo contains `a8s-deployment` as a submodule. However, using two separate `git clone` operations provides a more granular control over which version of each repository to checkout.
 
+* Create Postgres service instances with `a9s create pg instance`
+* Delete Postgres service instances with `a9s delete pg instance`
+
 # v0.7.0
 
 * Introduces a multi-k8s-provider support with implementations for both `minikube` and `kind`.